Rename classes and move methods for clarity
