+++ viewer-admin/src/main/java/nl/b3p/viewer/admin/stripes/ApplicationTreeLayerActionBean.java
@@ -27,11 +27,9 @@ import nl.b3p.viewer.config.ClobElement;
 import nl.b3p.viewer.config.app.*;
 import nl.b3p.viewer.config.security.Group;
 import nl.b3p.viewer.config.services.*;
-import nl.b3p.viewer.helpers.featuresources.FeatureSourceHelper;
 import nl.b3p.viewer.helpers.featuresources.SimpleFeatureTypeHelper;
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
 import nl.b3p.viewer.util.SelectedContentCache;
-import org.apache.commons.lang3.StringUtils;
 import org.json.*;
 import org.stripesstuff.stripersist.Stripersist;
 
@@ -245,7 +243,7 @@ public class ApplicationTreeLayerActionBean extends ApplicationActionBean {
             Layer layer = applicationLayer.getService().getSingleLayer(applicationLayer.getLayerName(), Stripersist.getEntityManager());
             if(layer != null && layer.getFeatureType() != null) {
                 SimpleFeatureType sft = layer.getFeatureType();
-                List<String> beh = FeatureSourceHelper.calculateUniqueValues(sft, attribute, null);
+                List<String> beh = FeatureSourceFactoryHelper.calculateUniqueValues(sft, attribute, null);
                 json.put("uniqueValues", new JSONArray(beh));
                 json.put("success", Boolean.TRUE);
             }
