+++ viewer-admin/src/test/java/nl/b3p/viewer/admin/stripes/WFSTypeNamingTest.java
@@ -27,7 +27,7 @@ import nl.b3p.viewer.config.services.GeoService;
 import nl.b3p.viewer.config.services.Layer;
 import nl.b3p.viewer.config.services.SimpleFeatureType;
 import nl.b3p.viewer.config.services.WFSFeatureSource;
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
 import nl.b3p.viewer.util.TestUtil;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -43,9 +43,6 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 import org.opengis.feature.Feature;
 
@@ -163,7 +160,7 @@ public class WFSTypeNamingTest extends TestUtil {
                     log.debug("Inspecting layer, name: " + lyr.getName() + ", featuretype: " + lyr.getFeatureType().getDescription());
                     assertTrue("Attribute type name does contain a colon", lyr.getFeatureType().getTypeName().contains(":"));
                     try {
-                        FeatureSource fs2 = SourceFactoryHelper.openGeoToolsFeatureSource(lyr.getFeatureType());
+                        FeatureSource fs2 = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(lyr.getFeatureType());
                         assertThat("No exception was thrown and featuresource not null", fs2, not(nullValue()));
                     } catch (Exception ex) {
                         log.error("Opening featuresource failed.", ex);
@@ -197,7 +194,7 @@ public class WFSTypeNamingTest extends TestUtil {
                 log.debug("Testing type: " + type.getTypeName());
                 assertTrue("Type name does contain a colon", type.getTypeName().contains(":"));
                 try {
-                    FeatureSource fs2 = SourceFactoryHelper.openGeoToolsFeatureSource(type);
+                    FeatureSource fs2 = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(type);
                     assertThat("No exception was thrown and featuresource not null", fs2, not(nullValue()));
                     
                     Query q = new Query(fs2.getName().toString());
