+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/helpers/services/GeoServiceHelper.java
@@ -7,7 +7,7 @@ import nl.b3p.web.WaitPageStatus;
 import javax.persistence.EntityManager;
 import java.util.Map;
 
-public interface ServiceHelper {
+public interface GeoServiceHelper {
 
     public GeoService loadServiceFromURL(String url, Map params, WaitPageStatus status, EntityManager em) throws Exception;
     public UpdateResult updateService(EntityManager em, GeoService service) throws Exception;
