+++ viewer/src/main/java/nl/b3p/viewer/stripes/AttributesActionBean.java
@@ -34,8 +34,8 @@ import nl.b3p.viewer.config.app.ApplicationLayer;
 import nl.b3p.viewer.config.app.ConfiguredAttribute;
 import nl.b3p.viewer.config.security.Authorizations;
 import nl.b3p.viewer.config.services.*;
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
-import nl.b3p.viewer.helpers.featuresources.WFSSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.WFSFeatureSourceHelper;
 import nl.b3p.viewer.util.ChangeMatchCase;
 import nl.b3p.viewer.util.FeatureToJson;
 import nl.b3p.viewer.util.FlamingoCQL;
@@ -501,7 +501,7 @@ public class AttributesActionBean extends LocalizableApplicationActionBean imple
                 if(isDebug() && ft.getFeatureSource() instanceof WFSFeatureSource) {
                     Map extraDataStoreParams = new HashMap();
                     extraDataStoreParams.put(WFSDataStoreFactory.TRY_GZIP.key, Boolean.FALSE);
-                    fs = WFSSourceHelper.openGeoToolsFSFeatureSource(ft, extraDataStoreParams, (WFSFeatureSource) ft.getFeatureSource());
+                    fs = WFSFeatureSourceHelper.openGeoToolsFSFeatureSource(ft, extraDataStoreParams, (WFSFeatureSource) ft.getFeatureSource());
                 } /*else if(ft.getFeatureSource() instanceof ArcGISFeatureSource) {
                     Map extraDataStoreParams = new HashMap();
                     if(isDebug()) {
@@ -513,7 +513,7 @@ public class AttributesActionBean extends LocalizableApplicationActionBean imple
                     fs = ((ArcGISFeatureSource)ft.getFeatureSource()).openGeoToolsFeatureSource(layer.getFeatureType(), extraDataStoreParams);
                 }*/ else {
 
-                    fs = SourceFactoryHelper.openGeoToolsFeatureSource(ft);
+                    fs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(ft);
                 }
 
                 boolean startIndexSupported = fs.getQueryCapabilities().isOffsetSupported();
