+++ viewer/src/main/java/nl/b3p/viewer/stripes/DataStoreSpinupActionBean.java
@@ -38,7 +38,7 @@ import nl.b3p.viewer.config.services.FeatureSource;
 import nl.b3p.viewer.config.services.GeoService;
 import nl.b3p.viewer.config.services.Layer;
 import nl.b3p.viewer.config.services.SimpleFeatureType;
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.geotools.feature.FeatureCollection;
@@ -196,7 +196,7 @@ public class DataStoreSpinupActionBean implements ActionBean {
 
                             org.geotools.data.FeatureSource gtfs = null;
                             try {
-                                gtfs = SourceFactoryHelper.openGeoToolsFeatureSource(sft);
+                                gtfs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(sft);
                                 FeatureCollection fc = gtfs.getFeatures(q);
                                 long time = System.currentTimeMillis() - startTime;
                                 log.debug(String.format("Type name \"%s\": request took %dms (feature count: %d)",
