+++ viewer/src/main/java/nl/b3p/viewer/stripes/FeatureInfoActionBean.java
@@ -17,7 +17,7 @@
 package nl.b3p.viewer.stripes;
 
 import net.sourceforge.stripes.controller.LifecycleStage;
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Point;
@@ -297,7 +297,7 @@ public class FeatureInfoActionBean extends LocalizableApplicationActionBean impl
                     }
                     String filter = query.optString("filter", null);
 
-                    fs = SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType(), TIMEOUT);
+                    fs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType(), TIMEOUT);
                     Query q = new Query(fs.getName().toString());
 
                     String geomAttribute = fs.getSchema().getGeometryDescriptor().getLocalName();
@@ -428,7 +428,7 @@ public class FeatureInfoActionBean extends LocalizableApplicationActionBean impl
                     }
 
                     String filter = query.optString("filter", null);
-                    fs = SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType(), TIMEOUT);
+                    fs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType(), TIMEOUT);
                     Query q = new Query(fs.getName().toString());
                     String geomAttribute = fs.getSchema().getGeometryDescriptor().getLocalName();
 
@@ -529,7 +529,7 @@ public class FeatureInfoActionBean extends LocalizableApplicationActionBean impl
                         q.setMaxFeatures(10 + 1);
                         q.setHandle("FeatureReportActionBean_related_attributes");
 
-                        fs = SourceFactoryHelper.openGeoToolsFeatureSource(fType, TIMEOUT);
+                        fs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(fType, TIMEOUT);
                         JSONArray features = ftjson.getJSONFeatures(al, fType, fs, q,em,application, context.getRequest());
 
                         JSONArray jsonFeats = new JSONArray();
