+++ viewer/src/main/java/nl/b3p/viewer/stripes/SplitFeatureActionBean.java
@@ -16,7 +16,7 @@
  */
 package nl.b3p.viewer.stripes;
 
-import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceFactoryHelper;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.LineString;
@@ -150,7 +150,7 @@ public class SplitFeatureActionBean extends LocalizableApplicationActionBean imp
                     throw new IllegalArgumentException(getBundle().getString("viewer.splitfeatureactionbean.2"));
                 }
 
-                fs = SourceFactoryHelper.openGeoToolsFeatureSource(this.layer.getFeatureType());
+                fs = FeatureSourceFactoryHelper.openGeoToolsFeatureSource(this.layer.getFeatureType());
                 if (!(fs instanceof SimpleFeatureStore)) {
                     throw new IllegalArgumentException(getBundle().getString("viewer.splitfeatureactionbean.3"));
                 }
